move publishing plots in pdf report to new file
